+++ src/models/Navigation.php
@@ -2,6 +2,7 @@
 namespace verbb\cpnav\models;
 
 use verbb\cpnav\CpNav;
+use verbb\cpnav\models\Settings;
 
 use Craft;
 use craft\base\Model;
@@ -48,6 +49,7 @@ class Navigation extends Model
     public ?string $customIcon = null;
     public ?string $type = null;
     public bool $newWindow = false;
+    public ?string $subnavBehaviour = null;
     public ?DateTime $dateCreated = null;
     public ?DateTime $dateUpdated = null;
     public ?string $uid = null;
@@ -201,6 +203,24 @@ class Navigation extends Model
         return '';
     }
 
+    public function getSubnavBehaviour(): ?string
+    {
+        if ($this->getChildren()) {
+            /* @var Settings $settings */
+            $settings = CpNav::$plugin->getSettings();
+
+            $behaviour = $settings->defaultSubnavBehaviour;
+
+            if ($this->subnavBehaviour) {
+                $behaviour = $this->subnavBehaviour;
+            }
+
+            return $behaviour;
+        }
+
+        return null;
+    }
+
     public function getLayout(): ?Layout
     {
         if ($this->_layout !== null) {
@@ -284,6 +304,11 @@ class Navigation extends Model
             }
 
             $children[] = $child;
+        }
+
+        // Return no children if the subnav is only set to show when active (and this isn't active)
+        if ($this->getSubnavBehaviour() === Settings::SUBNAV_DEFAULT && !$this->isSelected()) {
+            return [];
         }
 
         return $children;